float

+++ skeleton/kma.c
@@ -108,17 +108,17 @@ void fail();
 /**************Implementation***********************************************/
 
 //timing stuff
-double malloc_avg,free_avg,malloc_worst,free_worst;
-double cpu_time_free, cpu_time_malloc;
-double startFree, endFree,startMalloc,endMalloc;
-double freeRequests,mallocRequests;
+float malloc_avg,free_avg,malloc_worst,free_worst;
+float cpu_time_free, cpu_time_malloc;
+float startFree, endFree,startMalloc,endMalloc;
+float freeRequests,mallocRequests;
 
-freeRequests = 0;
-mallocRequests = 0;
-malloc_worst = 0;
-free_worst = 0;
-malloc_avg = 0;
-free_avg = 0;
+freeRequests = 0.0;
+mallocRequests = 0.0;
+malloc_worst = 0.0;
+free_worst = 0.0;
+malloc_avg = 0.0;
+free_avg = 0.0;
 
 int anyMismatches = 0;
 
@@ -308,7 +308,7 @@ allocate(mem_t* requests, int req_id, int req_size)
   startMalloc = clock();
   new->ptr = kma_malloc(new->size);
   endMalloc = clock();
-  cpu_time_malloc = ((double)(endMalloc - startMalloc)) / CLOCKS_PER_SEC;
+  cpu_time_malloc = ((float)(endMalloc - startMalloc)) / CLOCKS_PER_SEC;
   malloc_avg += cpu_time_malloc;
   if (cpu_time_malloc > malloc_worst)
   {
@@ -372,7 +372,7 @@ deallocate(mem_t* requests, int req_id)
   kma_free(cur->ptr, cur->size);
   endFree = clock();
   //add to time
-  cpu_time_free = ((double)(endFree - startFree)) / CLOCKS_PER_SEC;
+  cpu_time_free = ((float)(endFree - startFree)) / CLOCKS_PER_SEC;
   
   free_avg+= cpu_time_free;
   if (cpu_time_free > free_worst)
